Fix mod_load.c finetune conversion -fsanitize=shift-base

+++ src/loaders/mod_load.c
@@ -655,7 +655,7 @@ static int mod_load(struct module_data *m, HIO_HANDLE *f, const int start)
 	}
 	xxs->flg = (mh.ins[i].loop_size > 1 && xxs->lpe >= 4) ?
 		XMP_SAMPLE_LOOP : 0;
-	sub->fin = (int8)(mh.ins[i].finetune << 4);
+	sub->fin = (int8)((uint8)mh.ins[i].finetune << 4);
 	sub->vol = mh.ins[i].volume;
 	sub->pan = XMP_INST_NO_DEFAULT_PAN;
 	sub->sid = i;
